+++ backend/src/main/java/team/teamby/teambyteam/icalendar/application/IcalendarPublishService.java
@@ -5,9 +5,9 @@ import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import team.teamby.teambyteam.filesystem.FileCloudUploader;
-import team.teamby.teambyteam.icalendar.domain.IcalendarFileName;
+import team.teamby.teambyteam.icalendar.domain.vo.IcalendarFileName;
 import team.teamby.teambyteam.icalendar.domain.IcalendarParser;
-import team.teamby.teambyteam.icalendar.domain.PublishUrl;
+import team.teamby.teambyteam.icalendar.domain.vo.PublishUrl;
 import team.teamby.teambyteam.icalendar.domain.PublishedIcalendar;
 import team.teamby.teambyteam.icalendar.domain.PublishedIcalendarRepository;
 import team.teamby.teambyteam.schedule.domain.Schedule;
