+++ backend/src/main/java/team/teamby/teambyteam/icalendar/domain/PublishedIcalendar.java
@@ -10,6 +10,8 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import team.teamby.teambyteam.global.domain.BaseEntity;
+import team.teamby.teambyteam.icalendar.domain.vo.IcalendarFileName;
+import team.teamby.teambyteam.icalendar.domain.vo.PublishUrl;
 
 @Entity
 @Getter
